Extract ColocatedTemplateProcessor to dedicated file.
